use webpack in combination with concat
