+++ Gruntfile.js
@@ -71,6 +71,7 @@ module.exports = function(grunt) {
     'assemble',
     'handlebars',
     'modernizr',
+    'webpack',
     'concat',
     'sass:dev',
     'replace',
