+++ website-guts/assets/js/components/utility-nav.js
@@ -1,5 +1,5 @@
 var lastDropdown;
-var loggedInUtilityNav = require('../../../templates/client/loggedInUtilityNav');
+var loggedInUtilityNav = require('loggedInUtilityNav');
 
 function bindDropdownClick($dropdownMenus) {
