+++ website-guts/assets/js/utils/smooth_scroll.js
@@ -1,0 +1,10 @@
+window.optly.mrkt.utils.smoothScroll = function(event) {
+
+	var targetElmPos = $(this.getAttribute('href')).offset().top;
+
+	event.preventDefault();
+
+	$('html,body').animate({
+		scrollTop: targetElmPos
+	}, 1000);
+};
