+++ website-guts/assets/js/vendor.js
@@ -1,2 +1,5 @@
-require('jquery');
-require('oform');
+require('script!jquery-cookie/jquery.cookie');
+require('script!history.js');
+require('script!momentjs');
+require('script!fitvids');
+require('script!oform/src/oForm');
